chore: move get_shortnet_target and set_logger_env to common
